refactor: move domain interfaces to domain layer
