+++ app/src/main/java/br/com/jonatas/metronomeplus/data/provider/AssetProviderImpl.kt
@@ -2,6 +2,7 @@ package br.com.jonatas.metronomeplus.data.provider
 
 import android.content.Context
 import android.content.res.AssetManager
+import br.com.jonatas.metronomeplus.domain.provider.AssetProvider
 
 class AssetProviderImpl(private val context: Context) : AssetProvider {
     override fun getAssets(): AssetManager = context.assets
