+++ app/src/main/java/br/com/jonatas/metronomeplus/data/provider/AudioSettingProviderImpl.kt
@@ -3,6 +3,7 @@ package br.com.jonatas.metronomeplus.data.provider
 import android.content.Context
 import android.content.Context.AUDIO_SERVICE
 import android.media.AudioManager
+import br.com.jonatas.metronomeplus.domain.provider.AudioSettingsProvider
 
 class AudioSettingProviderImpl(private val context: Context) : AudioSettingsProvider {
     private val myAudioMgr = context.getSystemService(AUDIO_SERVICE) as AudioManager
