+++ app/src/main/java/br/com/jonatas/metronomeplus/domain/provider/AssetProvider.kt
@@ -1,4 +1,4 @@
-package br.com.jonatas.metronomeplus.data.provider
+package br.com.jonatas.metronomeplus.domain.provider
 
 import android.content.res.AssetManager
