+++ app/src/main/java/br/com/jonatas/metronomeplus/domain/provider/AudioSettingsProvider.kt
@@ -1,4 +1,4 @@
-package br.com.jonatas.metronomeplus.data.provider
+package br.com.jonatas.metronomeplus.domain.provider
 
 interface AudioSettingsProvider {
     fun getSampleRate(): Int
